Narrow softmax_chunked_kernel chunk count to Int32.
Softmax's chunked kernel was looping over `Int32(0):num_chunks - Int32(1)`
where `num_chunks` was Int64 (because the kernel takes `n_elems::Int`).
That widens the inner ForOp to `tile<i64>` and forces a `trunci` per
iteration in the offset arithmetic. Casting `num_chunks` to Int32 keeps
the loop in i32 throughout, matching cuTile Python's IR shape.

Bench (RTX 5080, 4096² Float32 chunked): 1587 → 1601 GB/s.

+++ examples/softmax.jl
@@ -51,7 +51,9 @@ function softmax_chunked_kernel(output::ct.TileArray{T,2}, input::ct.TileArray{T
     pid = ct.bid(1)
     num_programs = ct.num_blocks(1)
     M = size(input, 2)
-    num_chunks = (n_elems + TILE_SIZE - Int32(1)) ÷ Int32(TILE_SIZE)
+    # Narrow num_chunks to Int32 so the chunk loop stays in i32 (avoids
+    # i64↔i32 truncation on every offset; matches cuTile Python's IR).
+    num_chunks = Int32((n_elems + TILE_SIZE - Int32(1)) ÷ Int32(TILE_SIZE))
     row_offsets_base = ct.arange(TILE_SIZE)
 
     col_idx = pid
